requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
business-rules
